moves skip
